fs_mgr: add resize2fs call after e2fsck on "check" flag in fstab
In order to accomodate our method of disk image creation, statically linked
binaries for e2fsck and resize2fs were added to the ramdisk at the root
directory. If the "check" flag exists for an entry in the fstab, then both
"e2fsck -f -y <dev>" and "resize2fs <dev>" are performed. This process ensures
that a filesystem smaller than the partition it occupies will grow to use the
unclaimed space (as is the case for /data).

Signed-off-by: Robert Jones <rjones@gateworks.com>